chore: remove strict check on command arguments (#23927)
As part of #23907 I updated the dependency on `@types/yargs` to be in the `15` major version to reflect the version we use in the rest of the repo. Version `15` did not have the `strictOptions` yargs method so I attempted to enable `strict()` and disable `strictCommands` but it looks like that doesn't work. Removing strict altogether.

Our only other option is to keep `@types/yargs` pinned at `17.0.15` since `17.0.16` is no longer compatible with our version of typescript.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*