chore: pin and update dependencies (#24129)
Due to https://github.com/DefinitelyTyped/DefinitelyTyped/issues/64266, `@types/node` can't receive any minor updates or it will break us. So, we need to remove it from the minor upgrades allowed by the scripts. 

Fixing that set of dependencies required updates to some instances of `@types/jest` and `@types/fs-extra`. 

The rest of the dependency updates are as a result of running the script to make sure it behaved as expected.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*