feat(cdk): `cdk diff --quiet` suppresses progress messages (#26652)
Hello maintainers

This is my first PR, so all feedback is welcome. This change adds support for the `--quiet` flag within cdk diff command, for users that would not like the following messages 'Stack <name>' and 'There were no differences' to be silenced for stacks which do not have *any* identified differences. The request was made by @nomike in the #26526, I simply followed the very well detailed description.

The change follows similar structure to existing command and how they are passed, I have also added to the README as per the contributing guidelines.

Before submitting I have.
- Built the packages & linted them
- Ran unit tests for the entire module
- Ran all integration tests in the suite from `@aws-cdk-testing/cli-integ/bin/run-suite -a cli-integ-tests` 
- E2E tested by launching my small stack in account and executing diffs on it from my build.

Closes #26526

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*