fix(cli): hotswapping appsync functions fails when API does not return function on the first page (#31406)
### Issue # (if applicable)

https://github.com/aws-amplify/amplify-category-api/issues/2307

### Reason for this change



Hotswap for AppSync functions sometimes fails with `Deployment failed: MissingRequiredParameter: Missing required key 'functionId' in params`. This is because the SDK list function only retrieves the first page. In APIs where there are many functions it is possible for the hotswapped function to not be contained in the first page. This results in the previously mentioned error because the `functionId` is never retrieved.

### Description of changes



List all AppSync functions for a given API by iterating through the pages with the `nextToken`. The change was modeled after a similar request [here](https://github.com/aws/aws-cdk/blob/4371c93478594c50584642c4aae576f15ae6bfc2/packages/aws-cdk/lib/api/evaluate-cloudformation-template.ts#L23-L36).

### Description of how you validated changes



* Unit tests
* Integ tests
* Manual testing in app from https://github.com/aws-amplify/amplify-category-api/issues/2307

### Checklist
- [x] My code adheres to the [CONTRIBUTING GUIDE](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md) and [DESIGN GUIDELINES](https://github.com/aws/aws-cdk/blob/main/docs/DESIGN_GUIDELINES.md)

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*