feat(cli): cdk rollback (#31684)
This is a re-draft of #31407.  All description and motivation of the previous PR still apply.

The previous PR caused a regression because some `CREATE_IN_PROGRESS` events for CloudFormation do not have a `PhysicalResourceId`.

Fix that issue in this PR. Update the existing unit test that was supposed to catch this issue previously, it did not set a `ResourceStatus` which caused the event to be skipped for the wrong reason.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*